cli: bench: split sub commands into modules
